# crates/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "easyImmerse API",
    "description": "The REST API of easyImmerse.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/{id}/media": {
      "get": {
        "tags": [
          "media"
        ],
        "operationId": "listMediaFiles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The project id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project's media files, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListMediaFilesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Adds a media file to a project. A `path` source must name an existing file on the\nserver's machine, which only a token allowed to read local paths may do; the text\nsubtitle tracks inside it and the subtitle files beside it that share its name are added\nas its subtitle tracks.",
        "operationId": "addMediaFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The project id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMediaFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The added media file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaFile"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The token may not name local paths",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No such project, or no file at the given path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
    "/projects/{id}/media/{media_id}": {
      "delete": {
        "tags": [
          "media"
        ],
        "operationId": "removeMediaFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The project id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "media_id",
            "in": "path",
            "description": "The media file id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The media file was removed"
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No such media file in the project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddMediaFileRequest": {
        "type": "object",
        "required": [
          "name",
          "source"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name shown in the project's media list, usually the file name."
          },
          "source": {
            "$ref": "#/components/schemas/MediaFileSource"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable, machine-readable identifier such as `unauthorized`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ListMediaFilesResponse": {
        "type": "object",
        "required": [
          "media_files"
        ],
        "properties": {
          "media_files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaFile"
            }
          }
        }
      },
...
      "MediaFile": {
        "type": "object",
        "description": "A video or audio file added to a project.",
        "required": [
          "id",
          "project_id",
          "name",
          "source",
          "created_at_ms"
        ],
        "properties": {
          "created_at_ms": {
            "type": "integer",
            "format": "int64",
            "description": "Milliseconds since the Unix epoch.",
            "minimum": 0
          },
          "id": {
            "$ref": "#/components/schemas/MediaFileId"
          },
          "name": {
            "type": "string",
            "description": "The name shown in the project's media list, usually the file name."
          },
          "origin": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MediaOrigin",
                "description": "Where the file was fetched from, when a media-source plugin fetched it."
              },
              {
                "type": "null"
              }
            ]
          },
          "project_id": {
            "$ref": "#/components/schemas/ProjectId"
          },
          "source": {
            "$ref": "#/components/schemas/MediaFileSource"
          },
          "track_selection_json": {
            "type": [
              "string",
              "null"
            ],
            "description": "The user's saved choice of video and audio tracks, as the JSON the media crate\ndefines. Null until the user has chosen."
          }
        }
      },
      "MediaFileId": {
        "type": "string"
      },
      "MediaFileSource": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "path",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "path"
                ]
              },
              "path": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "size",
              "last_modified_ms",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "browser_file"
                ]
              },
              "last_modified_ms": {
                "type": "integer",
                "format": "int64",
                "description": "The file's modification time in milliseconds since the Unix epoch.",
                "minimum": 0
              },
              "size": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          }
        ],
        "description": "Where the bytes of a media file come from.\n\nThe `path` variant names a file on the machine running the server. The `browser_file`\nvariant describes a file the web app holds in memory; the server never sees its bytes,\nso it records only what identifies the file when the user picks it again."
      },
      "MediaLocator": {
        "type": "string",
        "description": "Identifies a media item at an external source, in whatever form the source's plugin\nunderstands, such as a URL."
      },
      "MediaOrigin": {
        "type": "object",
        "description": "The external source a media-source plugin fetched a media file from, kept so that\nthe same plugin can fetch more of what the source offers for it, such as subtitles.",
        "required": [
          "plugin",
          "locator"
        ],
        "properties": {
          "locator": {
            "$ref": "#/components/schemas/MediaLocator"
          },
          "plugin": {
            "type": "string",
            "description": "The name of the media-source plugin."
          }
        }
      },
...
      "ProjectId": {
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}